[tombstone] Renamed CLI to REPL
