+++ tombstone/src/main.rs
@@ -10,7 +10,7 @@ use crossterm::terminal::enable_raw_mode;
 use ratatui::Terminal;
 use ratatui::backend::CrosstermBackend;
 
-pub mod cli;
+pub mod repl;
 pub mod command;
 pub mod config;
 pub mod display_windows;
